Modify CppModel::ProjectInfo
Adding per project node information, to pass on the correct
defines/includes for each file, instead of aggregating them incorrectly.

Also split up SOURCES and OBJECTIVE_SOURCES.

Also ask the toolchain to convert the compilerflags to flags the
codemodel understands, for now only gcc and only c++11.

Also make the toolchain aware of the flags used to compile, so that it
can emit the correct defines.

Note: No header files are passed on.