Clang: Add synchronization code to the indexer
Better controll over the indexer current status, data synchronization
between the indexer and the indexer processor, which is the entity
that actually parses the translation units and collects the symbols.

The indexer is still not thread-safe, but now the processor is already
synchronized with it since it uses only local data and reports the
symbols through the future interface.