Clang: Improve sharing and life time control over units
More flexibility and clarity to control unit sharing. Now the units manager
acts as a container of "weak" units, being able to remove a particular one
in the case there are no more entities tracking it.

In addition, a few renames was done so the correspond more accurately
to what the functions really mean.

Change-Id: I5d329c0d1b015bde745e72b384669d9a6aa08829
Reviewed-by: Leandro Melo <leandro.melo@nokia.com>