Clang: Change index implementation
Move the previous generic database code into a more specific implementation,
now called index. This should allow support for the particular use-cases we
need in a simpler and more efficient way. Nevertheless, it might be that once
we have them well defined, the implementation could be lifted into a more
generic form again.

Change-Id: I887defaee875c8efb3742b02f25705116cbd4777
Reviewed-by: Leandro Melo <leandro.melo@nokia.com>