Clang: Introduce index serialization mechanism
Together with it there are other small fixes, some renaming,
a new class that performs simple #include tracking, and a
dependency graph implementation.

The work is not yet complete, there are still cases to handle
and issues to investigate. Also, a few improvements need to be made.
Nevertheless, this should establish the core principle.

Change-Id: Ibaa6fffafec1824131528284f8e92dba04d3b9cc
Reviewed-by: Christian Kamm <christian.d.kamm@nokia.com>